pyproject.toml
[project]
name = "shopdesk-support"
version = "0.1.0"
description = "an end-to-end automated customer support system powered by cutting edge AI technologies"
authors = [
    {name = "Oladimeji",email = "boladimeji834@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "fastapi (>=0.131.0,<0.132.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "faker (>=40.5.1,<41.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "pinecone (>=8.1.0,<9.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "email-validator (>=2.3.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
